requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-ckeditor==6.7.0
django-redis==5.2.0
redis==5.0.1
googletrans==4.0.0-rc1
pytest==7.4.0
pytest-django==4.5.2
flake8==6.1.0
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.7